moving validation error handler to helper lib
